refactor : 알림 아이콘 컴포넌트화
