+++ src/pages/auction/index.tsx
@@ -1,10 +1,12 @@
 import ArtWorkItem from '@components/auction/ArtWorkItem';
-import AuctionNavigate from '@components/auction/AuctionNavigate';
 import Layout from '@components/common/Layout';
+import Navigate from '@components/common/Navigate';
+import NoticeIcon from '@components/common/NoticeIcon';
 import Tab from '@components/common/Tab';
 import Image from 'next/image';
 import React from 'react';
 import tw from 'tailwind-styled-components';
+import { useRouter } from 'next/router';
 
 const ArtworkList = tw.div<DefaultProps>``;
 
@@ -13,10 +15,25 @@ interface DefaultProps {
 }
 
 export default function Auction() {
+  const router = useRouter();
+
   return (
     <>
       <Layout>
-        <AuctionNavigate />
+        <Navigate
+          left_message={
+            <Image
+              alt="logo"
+              src="/svg/icons/icon_logo.svg"
+              width="90"
+              height="0"
+            />
+          }
+          handleLeftButton={() => {
+            router.push('/home');
+          }}
+          right_message={<NoticeIcon />}
+        />
         <section className="relative mb-7 flex justify-between">
           <span className="text-[20px] font-bold">제 1회 아띠즈 경매</span>
           <div className=" flex items-center rounded border border-brand px-2">
