+++ src/pages/home/index.tsx
@@ -1,7 +1,6 @@
 import 'swiper/css';
 import 'swiper/css/pagination';
 
-import AuctionNavigate from '@components/auction/AuctionNavigate';
 import Layout from '@components/common/Layout';
 import Tab from '@components/common/Tab';
 import AuctionItem from '@components/home/AuctionItem';
@@ -19,7 +18,8 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import { isUser } from '@utils/isUser';
 import { Pagination } from 'swiper';
 import KeywordBox from '@components/common/KeywordBox';
-import useGetIsNotice from '@hooks/queries/useGetIsNotice';
+import Navigate from '@components/common/Navigate';
+import NoticeIcon from '@components/common/NoticeIcon';
 
 interface KeywordArtwork {
   id: string;
@@ -103,7 +103,21 @@ export default function Home() {
   return (
     <>
       <Layout>
-        <AuctionNavigate />
+        <Navigate
+          left_message={
+            <Image
+              alt="logo"
+              src="/svg/icons/icon_logo.svg"
+              width="90"
+              height="0"
+            />
+          }
+          handleLeftButton={() => {
+            router.push('/home');
+          }}
+          right_message={<NoticeIcon />}
+        />
+
         <section>
           <div className="text-14 text-[#767676]">
             {userInfo?.nickname}님 취향의
