+++ src/pages/profile/index.tsx
@@ -10,6 +10,7 @@ import Image from 'next/image';
 import tw from 'tailwind-styled-components';
 import { isUser } from '@utils/isUser';
 import { useRouter } from 'next/router';
+import NoticeIcon from '@components/common/NoticeIcon';
 
 interface defaultProps {
   [key: string]: any;
@@ -112,18 +113,17 @@ export default function Profile() {
     <>
       <Layout>
         <Navigate
-          left_message={false}
           message="프로필"
           isLeftButton={false}
+          handleLeftButton={() => {
+            router.push('/home');
+          }}
           right_message={
-            <Image
-              src="/svg/icons/icon_notification.svg"
-              alt="notification"
-              width="20"
-              height="0"
+            <NoticeIcon
+              isSearch={false}
+              handleRightButton={handleRightButton}
             />
           }
-          handleRightButton={handleRightButton}
         />
         <section>
           <WelcomeBox>
